BUILD: Make Python call more portable
We have to use a custom function in order to prefer finding a python
executable from PATH as this is the one that the user has most likely
installed the necessary packages for (this is definitely the case for
the CI runs). The fin_package version tends to find executables in very
exotic locations and does therefore seldom result in the expected
behavior.