FIX: Windows using different calling convention
It seems like the calling convention for the Windows API function `CloseHandle` can either vary between different
versions or at least varies in some circumstances. This would cause the compilation to fail because the calling convention
was mismatching with the one expected for the `handle_t` type.

This commit fixes this issue by using `decltype` to simply use the type that is used for the function, regardless what it might
be, instead of writing it down in advance.